Splitting CI between media and text
Just a refactoring. Two separate builds in CI, one for text and another one for media.

---------

+++ build.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Running only unit tests
-if [[ $TRAVIS_BRANCH != 'develop' && $TRAVIS_BRANCH != 'master' && ! $TRAVIS_COMMIT_MESSAGE =~ \[full\ ci\] && ! $TRAVIS_COMMIT_MESSAGE =~ \[smoke\ tests\] && ! $TRAVIS_COMMIT_MESSAGE =~ \[similarity\ tests\] ]]
+if [[ $TRAVIS_BRANCH != 'develop' && $TRAVIS_BRANCH != 'master' && ! $TRAVIS_COMMIT_MESSAGE =~ \[full\ ci\] && ! $TRAVIS_COMMIT_MESSAGE =~ \[smoke\ tests\] && ! $TRAVIS_COMMIT_MESSAGE =~ \[text\ similarity\ tests\] && ! $TRAVIS_COMMIT_MESSAGE =~ \[media\ similarity\ tests\] ]]
 then
   echo "Running only unit tests"
   docker-compose build web
@@ -14,34 +14,40 @@ else
     docker-compose build web api api-background pender pender-background
     docker-compose -f docker-compose.yml -f docker-test.yml up -d web api api-background pender pender-background chromedriver
   else
-    i=0
-    NGROK_URL=""
-    ngrok config add-authtoken $NGROK_AUTH
-    ngrok config upgrade 2
-    while [ -z "$NGROK_URL" -a $i -lt 5 ]; do
-      i=$(($i + 1))
-      ngrok http 9000 >/dev/null &
-      until curl --silent -I -f --fail http://localhost:4040; do printf "."; sleep 10; done
-      curl -I -v http://localhost:4040
-      curl localhost:4040/api/tunnels > ngrok.json
-      cat ngrok.json
-      NGROK_URL=$(grep -Po '"public_url": *\K"[^"]*"' ngrok.json | tail -n1 | sed 's/.\(.*\)/\1/' | sed 's/\(.*\)./\1/')
+    if [[ $TRAVIS_JOB_NAME == 'media-similarity-tests' ]]
+    then
+      i=0
+      NGROK_URL=""
+      ngrok config add-authtoken $NGROK_AUTH
+      ngrok config upgrade 2
+      while [ -z "$NGROK_URL" -a $i -lt 5 ]; do
+        i=$(($i + 1))
+        ngrok http 9000 >/dev/null &
+        until curl --silent -I -f --fail http://localhost:4040; do printf "."; sleep 10; done
+        curl -I -v http://localhost:4040
+        curl localhost:4040/api/tunnels > ngrok.json
+        cat ngrok.json
+        NGROK_URL=$(grep -Po '"public_url": *\K"[^"]*"' ngrok.json | tail -n1 | sed 's/.\(.*\)/\1/' | sed 's/\(.*\)./\1/')
+        if [ -z $NGROK_URL ]
+        then
+          kill -9 $(pgrep ngrok)
+        fi
+        sleep 5
+      done
       if [ -z $NGROK_URL ]
       then
-        kill -9 $(pgrep ngrok)
+        echo "Not able to connect a Ngrok Tunnel. Please try again!"
+        exit 1
       fi
-      sleep 5
-    done
-    if [ -z $NGROK_URL ]
-    then
-      echo "Not able to connect a Ngrok Tunnel. Please try again!"
-      exit 1
+      echo "Ngrok tunnel: $NGROK_URL"
+      sed -i "s~similarity_media_file_url_host: ''~similarity_media_file_url_host: '$NGROK_URL'~g" check-api/config/config.yml
+      cat check-api/config/config.yml | grep similarity_media_file_url_host
+      docker-compose build web api api-background pender pender-background chromedriver alegre presto-server presto-audio presto-image presto-video
+      docker-compose -f docker-compose.yml -f docker-test.yml up -d web api api-background pender pender-background chromedriver alegre presto-server presto-audio presto-image presto-video
+    else
+      docker-compose build web api api-background pender pender-background chromedriver alegre presto-server presto-mean-tokens
+      docker-compose -f docker-compose.yml -f docker-test.yml up -d web api api-background pender pender-background chromedriver alegre presto-server presto-mean-tokens
     fi
-    echo "Ngrok tunnel: $NGROK_URL"
-    sed -i "s~similarity_media_file_url_host: ''~similarity_media_file_url_host: '$NGROK_URL'~g" check-api/config/config.yml
-    cat check-api/config/config.yml | grep similarity_media_file_url_host
-    docker-compose build
-    docker-compose -f docker-compose.yml -f docker-test.yml up -d
     until curl --silent -I -f --fail http://localhost:3100; do printf .; sleep 1; done
     until curl --silent -I -f --fail http://localhost:8000/ping; do printf .; sleep 1; done
   fi
